+++ PIL/PngImagePlugin.py
@@ -35,10 +35,13 @@ from __future__ import print_function
 
 __version__ = "0.9"
 
+import logging
 import re
+import zlib
 
 from PIL import Image, ImageFile, ImagePalette, _binary
-import zlib
+
+logger = logging.getLogger(__name__)
 
 i8 = _binary.i8
 i16 = _binary.i16be
@@ -71,6 +74,7 @@ _MODES = {
 
 
 _simple_palette = re.compile(b'^\xff+\x00\xff*$')
+_null_palette = re.compile(b'^\x00*$')
 
 # Maximum decompressed size for a iTXt or zTXt chunk.
 # Eliminates decompression bombs where compressed chunks can expand 1000x
@@ -90,7 +94,7 @@ def _safe_zlib_decompress(s):
 # --------------------------------------------------------------------
 # Support classes.  Suitable for PNG and related formats like MNG etc.
 
-class ChunkStream:
+class ChunkStream(object):
 
     def __init__(self, fp):
 
@@ -128,8 +132,7 @@ class ChunkStream:
     def call(self, cid, pos, length):
         "Call the appropriate chunk handler"
 
-        if Image.DEBUG:
-            print("STREAM", cid, pos, length)
+        logger.debug("STREAM %s %s %s", cid, pos, length)
         return getattr(self, "chunk_" + cid.decode('ascii'))(pos, length)
 
     def crc(self, cid, data):
@@ -183,7 +186,7 @@ class iTXt(str):
         return self
 
 
-class PngInfo:
+class PngInfo(object):
     """
     PNG chunk container (for use with save(pnginfo=))
 
@@ -292,9 +295,8 @@ class PngStream(ChunkStream):
         # Compression method    1 byte (0)
         # Compressed profile    n bytes (zlib with deflate compression)
         i = s.find(b"\0")
-        if Image.DEBUG:
-            print("iCCP profile name", s[:i])
-            print("Compression method", i8(s[i]))
+        logger.debug("iCCP profile name %s", s[:i])
+        logger.debug("Compression method %s", i8(s[i]))
         comp_method = i8(s[i])
         if comp_method != 0:
             raise SyntaxError("Unknown compression method %s in iCCP chunk" %
@@ -350,6 +352,8 @@ class PngStream(ChunkStream):
                 i = s.find(b"\0")
                 if i >= 0:
                     self.im_info["transparency"] = i
+            elif _null_palette.match(s):
+                self.im_info["transparency"] = 0
             else:
                 self.im_info["transparency"] = s
         elif self.im_mode == "L":
@@ -504,8 +508,7 @@ class PngImageFile(ImageFile.ImageFile):
             except EOFError:
                 break
             except AttributeError:
-                if Image.DEBUG:
-                    print(cid, pos, length, "(unknown)")
+                logger.debug("%s %s %s (unknown)", cid, pos, length)
                 s = ImageFile._safe_read(self.fp, length)
 
             self.png.crc(cid, s)
@@ -620,7 +623,7 @@ def putchunk(fp, cid, *data):
     fp.write(o16(hi) + o16(lo))
 
 
-class _idat:
+class _idat(object):
     # wrap output from the encoder in IDAT chunks
 
     def __init__(self, fp, chunk):
@@ -771,7 +774,7 @@ def _save(im, fp, filename, chunk=putchunk, check=0):
 def getchunks(im, **params):
     """Return a list of PNG chunks representing this image."""
 
-    class collector:
+    class collector(object):
         data = []
 
         def write(self, data):