+++ Tests/helper.py
@@ -234,6 +234,9 @@ def imagemagick_available():
     return IMCONVERT and command_succeeds([IMCONVERT, '-version'])
 
 
+def on_appveyor():
+    return 'APPVEYOR' in os.environ
+
 if sys.platform == 'win32':
     IMCONVERT = os.environ.get('MAGICK_HOME', '')
     if IMCONVERT: