+++ Tests/test_file_gif.py
@@ -25,6 +25,12 @@ class TestFileGif(PillowTestCase):
         self.assertEqual(im.size, (128, 128))
         self.assertEqual(im.format, "GIF")
 
+    def test_invalid_file(self):
+        invalid_file = "Tests/images/flower.jpg"
+
+        self.assertRaises(SyntaxError,
+                          lambda: GifImagePlugin.GifImageFile(invalid_file))
+
     def test_optimize(self):
         from io import BytesIO
 
@@ -71,6 +77,24 @@ class TestFileGif(PillowTestCase):
 
         self.assert_image_similar(reread.convert('RGB'), hopper(), 50)
 
+    def test_roundtrip_save_all(self):
+        # Single frame image
+        out = self.tempfile('temp.gif')
+        im = hopper()
+        im.save(out, save_all=True)
+        reread = Image.open(out)
+
+        self.assert_image_similar(reread.convert('RGB'), im, 50)
+
+        # Multiframe image
+        im = Image.open("Tests/images/dispose_bgnd.gif")
+
+        out = self.tempfile('temp.gif')
+        im.save(out, save_all=True)
+        reread = Image.open(out)
+
+        self.assertEqual(im.n_frames, 5)
+
     def test_palette_handling(self):
         # see https://github.com/python-pillow/Pillow/issues/513
 
@@ -92,20 +116,20 @@ class TestFileGif(PillowTestCase):
 
         def roundtrip(im, *args, **kwargs):
             out = self.tempfile('temp.gif')
-            im.save(out, *args, **kwargs)
+            im.copy().save(out, *args, **kwargs)
             reloaded = Image.open(out)
 
-            return [im, reloaded]
+            return reloaded
 
         orig = "Tests/images/test.colors.gif"
         im = Image.open(orig)
 
-        self.assert_image_equal(*roundtrip(im))
-        self.assert_image_equal(*roundtrip(im, optimize=True))
+        self.assert_image_similar(im, roundtrip(im), 1)
+        self.assert_image_similar(im, roundtrip(im, optimize=True), 1)
 
         im = im.convert("RGB")
         # check automatic P conversion
-        reloaded = roundtrip(im)[1].convert('RGB')
+        reloaded = roundtrip(im).convert('RGB')
         self.assert_image_equal(im, reloaded)
 
     @unittest.skipUnless(netpbm_available(), "netpbm not available")
@@ -133,6 +157,27 @@ class TestFileGif(PillowTestCase):
                 img.seek(img.tell() + 1)
         except EOFError:
             self.assertEqual(framecount, 5)
+
+    def test_n_frames(self):
+        im = Image.open(TEST_GIF)
+        self.assertEqual(im.n_frames, 1)
+        self.assertFalse(im.is_animated)
+
+        im = Image.open("Tests/images/iss634.gif")
+        self.assertEqual(im.n_frames, 42)
+        self.assertTrue(im.is_animated)
+
+    def test_eoferror(self):
+        im = Image.open(TEST_GIF)
+
+        n_frames = im.n_frames
+        while True:
+            n_frames -= 1
+            try:
+                im.seek(n_frames)
+                break
+            except EOFError:
+                self.assertTrue(im.tell() < n_frames)
 
     def test_dispose_none(self):
         img = Image.open("Tests/images/dispose_none.gif")
@@ -169,6 +214,42 @@ class TestFileGif(PillowTestCase):
         # first frame
         self.assertEqual(img.histogram()[img.info['transparency']], 0)
 
+    def test_duration(self):
+        duration = 1000
+
+        out = self.tempfile('temp.gif')
+        fp = open(out, "wb")
+        im = Image.new('L', (100, 100), '#000')
+        for s in GifImagePlugin.getheader(im)[0] + GifImagePlugin.getdata(im, duration=duration):
+            fp.write(s)
+        fp.write(b";")
+        fp.close()
+        reread = Image.open(out)
+
+        self.assertEqual(reread.info['duration'], duration)
+
+    def test_number_of_loops(self):
+        number_of_loops = 2
+
+        out = self.tempfile('temp.gif')
+        fp = open(out, "wb")
+        im = Image.new('L', (100, 100), '#000')
+        for s in GifImagePlugin.getheader(im)[0] + GifImagePlugin.getdata(im, loop=number_of_loops):
+            fp.write(s)
+        fp.write(b";")
+        fp.close()
+        reread = Image.open(out)
+
+        self.assertEqual(reread.info['loop'], number_of_loops)
+
+    def test_background(self):
+        out = self.tempfile('temp.gif')
+        im = Image.new('L', (100, 100), '#000')
+        im.info['background'] = 1
+        im.save(out)
+        reread = Image.open(out)
+
+        self.assertEqual(reread.info['background'], im.info['background'])
 
 if __name__ == '__main__':
     unittest.main()