+++ Tests/test_file_mpo.py
@@ -95,6 +95,23 @@ class TestFileMpo(PillowTestCase):
             im.seek(0)
             self.assertEqual(im.tell(), 0)
 
+    def test_n_frames(self):
+        im = Image.open("Tests/images/sugarshack.mpo")
+        self.assertEqual(im.n_frames, 2)
+        self.assertTrue(im.is_animated)
+
+    def test_eoferror(self):
+        im = Image.open("Tests/images/sugarshack.mpo")
+
+        n_frames = im.n_frames
+        while True:
+            n_frames -= 1
+            try:
+                im.seek(n_frames)
+                break
+            except EOFError:
+                self.assertTrue(im.tell() < n_frames)
+
     def test_image_grab(self):
         for test_file in test_files:
             im = Image.open(test_file)